# userappAN/src/openapi/project.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8092/project",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/project/updateProject/{idProject}": {
      "put": {
        "tags": [
          "project-controller"
        ],
        "operationId": "updateProject",
        "parameters": [
          {
            "name": "idProject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/project/predict-status": {
      "post": {
        "tags": [
          "project-controller"
        ],
        "operationId": "predictProjectStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/project/addProject": {
      "post": {
        "tags": [
          "project-controller"
        ],
        "operationId": "addProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/project/statisticsByStatus": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "getStatisticsByStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/project/progress/{idProject}": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "getProjectProgress",
        "parameters": [
          {
            "name": "idProject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/project/getProjectsByStatus/{status}": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "countProjectsByStatus",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ON_GOING",
                "COMPLETED",
                "DELAYED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/project/getProjectById/{idProject}": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "findProjectById",
        "parameters": [
          {
            "name": "idProject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/project/getAllProjects": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "getAllProjects",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/project/geoAndRemaining/{idProject}": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "getGeoAndRemaining",
        "parameters": [
          {
            "name": "idProject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/project/averageDuration": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "getAverageDuration",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      }
    },
    "/project/deleteProject/{idProject}": {
      "delete": {
        "tags": [
          "project-controller"
        ],
        "operationId": "deleteProject",
        "parameters": [
          {
            "name": "idProject",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "idProject": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "ON_GOING",
              "COMPLETED",
              "DELAYED"
            ]
          },
          "city": {
            "type": "string"
          },
          "latitude": {
            "type": "string"
          },
          "longitude": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        }
      }
    }
  }
}